fix: address AuthZed review feedback
