+++ apps/web/lib/authzed/__mocks__/logger.ts
@@ -1,0 +1,10 @@
+import { vi } from "vitest";
+
+export const loggerMocks = {
+  debug: vi.fn(),
+  warn: vi.fn(),
+};
+
+vi.mock("@formbricks/logger", () => ({
+  logger: loggerMocks,
+}));
