+++ docker/__tests__/formbricks-script.test.ts
@@ -1,5 +1,5 @@
 import { execFileSync } from "node:child_process";
-import { existsSync, mkdtempSync, readFileSync, rmSync, writeFileSync } from "node:fs";
+import { existsSync, mkdtempSync, readFileSync, rmSync, statSync, writeFileSync } from "node:fs";
 import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { fileURLToPath } from "node:url";
@@ -15,6 +15,17 @@ const createTempDir = (): string => {
   tempDirs.push(tempDir);
   return tempDir;
 };
+
+const parseEnvFile = (contents: string): Map<string, string> =>
+  new Map(
+    contents
+      .trim()
+      .split("\n")
+      .map((line) => {
+        const separatorIndex = line.indexOf("=");
+        return [line.slice(0, separatorIndex), line.slice(separatorIndex + 1)];
+      })
+  );
 
 const addFormbricksTraefikLabels = (
   composePath: string,
@@ -60,6 +71,56 @@ afterEach(() => {
   for (const tempDir of tempDirs.splice(0)) {
     rmSync(tempDir, { recursive: true, force: true });
   }
+});
+
+describe("docker/formbricks.sh AuthZed setup", () => {
+  test("writes AuthZed secrets without printing them", () => {
+    const envPath = join(createTempDir(), ".env");
+    const authzedToken = "authzed-token-value";
+    const authzedDatabasePassword = "authzed-database-password";
+
+    const output = execFileSync(
+      "bash",
+      [
+        "-lc",
+        'source "$1"; write_base_env_file "$2" hub-key cube-secret "$3" "$4"',
+        "bash",
+        formbricksScriptPath,
+        envPath,
+        authzedToken,
+        authzedDatabasePassword,
+      ],
+      { encoding: "utf8" }
+    );
+
+    const env = parseEnvFile(readFileSync(envPath, "utf8"));
+
+    expect(output).toBe("");
+    expect(output).not.toContain(authzedToken);
+    expect(output).not.toContain(authzedDatabasePassword);
+    expect(env.get("AUTHZED_TOKEN")).toBe(authzedToken);
+    expect(env.get("AUTHZED_DATABASE_PASSWORD")).toBe(authzedDatabasePassword);
+    expect(statSync(envPath).mode & 0o777).toBe(0o600);
+  });
+
+  test("pins and verifies the downloaded bootstrap helper before making it executable", () => {
+    const script = readFileSync(formbricksScriptPath, "utf8");
+    const downloadStart = script.indexOf(
+      'authzed_bootstrap_commit="10d5ad908491a8a818aef3c6ada91fa4fdc30b03"'
+    );
+    const checksumStart = script.indexOf("sha256sum --check --status -", downloadStart);
+    const chmodStart = script.indexOf("chmod 700 authzed-postgres-bootstrap.sh", checksumStart);
+
+    expect(downloadStart).toBeGreaterThanOrEqual(0);
+    expect(script.slice(downloadStart, checksumStart)).not.toContain(
+      "formbricks/stable/docker/authzed-postgres-bootstrap.sh"
+    );
+    expect(script.slice(downloadStart, checksumStart)).toContain(
+      "70975701cdf0dcffef5d3573a7514360e87428bb07cc4bfb4dbf47ae0c2e93a5"
+    );
+    expect(checksumStart).toBeGreaterThan(downloadStart);
+    expect(chmodStart).toBeGreaterThan(checksumStart);
+  });
 });
 
 describe("docker/formbricks.sh Traefik label injection", () => {
